ConcurrencyManagedCDIBeans moved to cdi package and cdi package exported by osgi
