+++ src/main/java/org/glassfish/concurro/cdi/ConcurrencyManagedCDIBeans.java
@@ -13,7 +13,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
-package org.glassfish.concurro.internal;
+package org.glassfish.concurro.cdi;
 
 import java.util.HashMap;
 import java.util.Map;
